io_uring: widen zcrx_area_size to size_t
ABI was unsigned int — 32 bit on x86_64 — so EVPL_ZCRX_AREA_SIZE >= 4 GiB
silently wrapped to 0 and fell back to the 256 MiB default. The
underlying area_bytes variable in zcrx_setup is already size_t; the
config field and setter were the narrow link.

+++ include/evpl/evpl_config.h
@@ -5,6 +5,7 @@
 #pragma once
 
 #include <stdint.h>
+#include <stddef.h>
 
 #ifndef EVPL_INCLUDED
 #error "Do not include evpl_config.h directly, include evpl/evpl.h instead"
@@ -151,7 +152,7 @@ void evpl_global_config_set_io_uring_zcrx_rxq_count(
 
 void evpl_global_config_set_io_uring_zcrx_area_size(
     struct evpl_global_config *config,
-    unsigned int               size);
+    size_t                     size);
 
 void evpl_global_config_set_io_uring_zcrx_rq_entries(
     struct evpl_global_config *config,
